Refactor: Move column methods to `columns`, get_column_types testable
Loses one optimization:
The list of column values is now always created, not only when
necessary.
